[IA-66] WIP

+++ app/src/main/java/co/netguru/android/inbbbox/localrepository/database/GuestModeLikesRepository.java
@@ -8,6 +8,7 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import co.netguru.android.inbbbox.model.localrepository.database.DaoSession;
+import co.netguru.android.inbbbox.model.localrepository.database.ShotDB;
 import co.netguru.android.inbbbox.model.localrepository.database.ShotDBDao;
 import co.netguru.android.inbbbox.model.localrepository.database.mapper.ShotDBMapper;
 import co.netguru.android.inbbbox.model.localrepository.database.mapper.TeamDBMapper;
@@ -31,7 +32,7 @@ public class GuestModeLikesRepository {
     public Completable addLikedShot(@NonNull Shot shot) {
         return daoSession.rxTx()
                 .run(() -> {
-                    daoSession.getShotDBDao().insertOrReplace(ShotDBMapper.fromShot(shot));
+                    daoSession.getShotDBDao().insertOrReplace(ShotDBMapper.fromShot(increaseShotLikesCount(shot)));
                     daoSession.getUserDBDao().insertOrReplace(UserDBMapper.fromUser(shot.author()));
                     if (shot.team() != null) {
                         daoSession.getTeamDBDao().insertOrReplace(TeamDBMapper.fromTeam(shot.team()));
@@ -41,11 +42,18 @@ public class GuestModeLikesRepository {
     }
 
     public Observable<List<Shot>> getLikedShots() {
-        return daoSession.getShotDBDao().queryBuilder().rx().oneByOne().map(Shot::fromDB).toList();
+        return daoSession.getShotDBDao().queryBuilder().rx().oneByOne()
+                .filter(ShotDB::getIsLiked)
+                .map(Shot::fromDB)
+                .toList();
     }
 
     public Completable removeLikedShot(Shot shot) {
-        return daoSession.getShotDBDao().rx().deleteByKey(shot.id()).toCompletable();
+        if (!shot.isBucketed()) {
+            return daoSession.getShotDBDao().rx().deleteByKey(shot.id()).toCompletable();
+        }
+        return daoSession.getShotDBDao().rx()
+                .insertOrReplace(ShotDBMapper.fromShot(unlikeShot(shot))).toCompletable();
     }
 
     public Completable isShotLiked(Shot shot) {
@@ -54,11 +62,24 @@ public class GuestModeLikesRepository {
                 .rx()
                 .unique()
                 .flatMap(shotDB -> {
-                    if (shotDB == null) {
+                    if (shotDB == null || !shotDB.getIsLiked()) {
                         return Observable.error(new Throwable(SHOT_IS_NOT_LIKED_ERROR));
                     }
                     return Observable.empty();
                 })
                 .toCompletable();
+    }
+
+    private Shot increaseShotLikesCount(Shot shot) {
+        return Shot.update(shot)
+                .likesCount(shot.likesCount() + 1)
+                .build();
+    }
+
+    private Shot unlikeShot(Shot shot) {
+        return Shot.update(shot)
+                .likesCount(shot.likesCount() - 1)
+                .isLiked(false)
+                .build();
     }
 }
